Fix weird bug with mass coming out as -1 for ships

+++ src/LibreLancer.Data/Ships/Ship.cs
@@ -42,7 +42,7 @@ namespace LibreLancer.Data.Ships
         [Entry("hold_size")]
 		public int HoldSize;
         [Entry("mass")]
-		public int Mass;
+		public float Mass;
         [Entry("ship_class")]
 		public int ShipClass;
         [Entry("type")]
